refactored components. added InputSecure comp. Join page now has inputs
