+++ src/screens/JoinScreen.tsx
@@ -1,20 +1,63 @@
-import React, { memo } from "react";
-import { Text } from "react-native"
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import React, { memo } from 'react';
+import {
+  StyleSheet,
+  View,
+  Image,
+} from "react-native";
+import {
+  Button,
+} from "react-native-paper"
 
-import { Navigation } from "../types";
-import Background from "../components/Background";
-
-type Props = {
-    navigation: Navigation;
-};
+import Input from "../components/Input";
+import InputSecure from "../components/InputSecure"
+import { theme } from "../core/theme";
 
 const JoinScreen = () => {
-    
+
     return (
-        <Background>
-            <Text>Hello Join!</Text>
-        </Background>
-)};
+    <>
+    <Image source={require('../assets/logo-light.png')} style={styles.logo}/>
+    <View style={styles.inputContainer}>
+      <Input label="Firstname"></Input>
+      <Input label="Lastname"></Input>
+      <Input label="Email"></Input>
+      <Input label="Username"></Input>
+      <InputSecure label="Password"></InputSecure>
+      <View style={styles.inputOptionsContainer}></View>
+
+      <Button
+        mode="contained"
+        onPress={() => { console.log("Log In clicked"); }}
+        style={{alignSelf: 'center', width: 200, marginTop: 75}}
+      >
+        Log In
+      </Button>
+    </View>
+    </>
+  )
+};
+
+const styles = StyleSheet.create({
+  logo: {
+    flex: 1,
+    backgroundColor: theme.colors.background,
+    width: null,
+    height: null,
+    resizeMode: 'contain',
+    paddingTop: 50,
+  },
+  inputContainer: {
+    backgroundColor: theme.colors.background,
+    flex: 2,
+    justifyContent: 'flex-start',
+    paddingHorizontal: 25,
+  },
+  inputOptionsContainer: {
+    display: 'flex',
+    alignItems: "center",
+    justifyContent: 'center',
+    flexDirection: 'row',
+  },
+});
 
 export default memo(JoinScreen);
